Introduce ReactiveMongoSessionRepository.
